+++ package.json
@@ -14,6 +14,7 @@
     "@supabase/supabase-js": "^2.44.4",
     "@tanstack/react-query": "^5.51.11",
     "@tanstack/react-query-devtools": "^5.51.11",
+    "@team-mintc/portone-v2": "^0.7.9",
     "@types/react-modal": "^3.16.3",
     "axios": "^1.7.2",
     "next": "14.2.5",
